Add gallery plugin
Loads images from a static folder and selects a random image on plugin invocation.
The plugin is similar to the existing image upload
but different as it has no notion of different instance.
The idea is to have something a bit more user friendly with the drag and drop, preview and click to remove